actuary/requirements.txt
numpy==1.21.0
matplotlib==3.4.2
pandas==1.1.3
seaborn==0.11.1
streamlit==0.84.1
scikit-learn==0.24.2
auto-sklearn==0.14.2
scipy==1.10.1